% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreData.R
\name{exploreData}
\alias{exploreData}
\title{explore dataset for unusual values}
\usage{
exploreData(X, dp = -1, alpha = 0.1)
}
\arguments{
\item{X}{A vector/matrix/dataframe of data.}

\item{dp}{decimal places for returned statistics. Negative value implies no rounding. Defaults to -1.}

\item{alpha}{proportion for high and low quantiles. Defaults to 0.1}
}
\value{
A list of values for scaling the datasets.
}
\description{
explore dataset for unusual values
}
\examples{
exploreData(X)

}
