% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.R
\name{lin}
\alias{lin}
\title{number of inputs to each node}
\usage{
lin(design)
}
\arguments{
\item{design}{the coded design}
}
\value{
A vector containing the number of inputs to each node.
}
\description{
number of inputs to each node
}
\examples{
lin(myDesign)

}
