% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnDesign.R
\name{nnDesign}
\alias{nnDesign}
\title{create the design file for a neural network}
\usage{
nnDesign(architecture, actFun, lossFun, drop = NULL, skip = NULL, seed = -1)
}
\arguments{
\item{architecture}{vector containing the number of nodes in each layer}

\item{actFun}{string vector containing the activation functions for layers 2, 3 etc.}

\item{lossFun}{string containing the name of the loss function}

\item{seed}{optional positive seed for the initial values}
}
\value{
The encoded network design.
}
\description{
create the design file for a neural network
}
\examples{
nnDesign(arch=c(1, 4, 1), act=c("sigmoid", "identity"), loss="L2" )

}
