% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnPredict.R
\name{nnPredict}
\alias{nnPredict}
\title{predict responses for a fitted neural network}
\usage{
nnPredict(X, design)
}
\arguments{
\item{X}{A matrix of features.}

\item{design}{encoded version of the network design as created by nnDesign}
}
\value{
A matrix containing the predictions.
}
\description{
predict responses for a fitted neural network
}
\examples{
YT <- nnPredict(XT, design)

}
