% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleData.R
\name{scaleData}
\alias{scaleData}
\title{scale a dataset}
\usage{
scaleData(X, Y = NULL, stats)
}
\arguments{
\item{X}{A matrix/dataframe of data.}

\item{Y}{Optionally a second matrix/dataframe of data. Defaults to no second dataset.}

\item{stats}{list of scaling statistics as returned by function scaleStats}
}
\value{
A scaled copy of the dataset
}
\description{
scale a dataset
}
\examples{
sts <- scaleStats(X)
XS  <- scaleData(X, stats=sts)

}
